moving code from helpers init to helpers urls, changing imports, first name modification in route
